rm_pm_hal : fix CWARN.BITOP.SIZE
Operands in bitwise operation have diffrent size.
Fix made operands of same size in bitwise operation.

Fixes: SYSFW-6265

+++ pm/drivers/device/device.c
@@ -52,7 +52,7 @@ static sbool devices_rw;
 
 void pm_devgroup_set_enabled(devgrp_t groups)
 {
-	pm_devgroups_enabled |= groups;
+	pm_devgroups_enabled |= (u32)groups;
 }
 
 sbool pm_devgroup_is_enabled(devgrp_t groups)
